+++ internal/server/webservice/handlers/testdata/TestUpload/golden/empty_app_name
