+++ internal/server/webservice/handlers/testdata/TestUpload/golden/file_too_large
