+++ internal/server/webservice/handlers/testdata/TestUpload/golden/invalid_json
