+++ internal/server/webservice/handlers/testdata/TestUpload/golden/invalid_method_-_get
