+++ internal/server/webservice/handlers/testdata/TestUpload/golden/valid_upload
