+++ internal/server/webservice/testdata/TestServeMulti/golden/valid_upload_accepted
